LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-video-element/resize-during-playback-expected.txt


FAIL mp4 video assert_equals: Expected resize event, but got ended event instead expected "resize" but got "ended"
NOTRUN webm video webm supported
